+++ include/ishaders.h
@@ -59,7 +59,7 @@ enum
 };
 
 template<typename Element> class BasicVector3;
-typedef BasicVector3<double> Vector3;
+typedef BasicVector3<float> Vector3;
 typedef Vector3 Colour3;
 
 /**
